requirements.txt
django==6.0.1
psycopg2-binary
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
Pillow==11.0.0
